+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1 (February 11, 2016)
+
+- Updated verify payment
+- Added custom SimplePay checkout description
+- Updated plugin settings page
+
 ## 1.0.0 (January 29, 2016)
 
 - Initial release